test/playback/requirements.txt
numpy==1.20.1
paho-mqtt==1.5.0